move SlugGenerator to Services namespace, and refactor tests to its own class
