+++ BlogTemplate/Pages/New.cshtml.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using BlogTemplate.Models;
+using BlogTemplate.Services;
 
 namespace BlogTemplate.Pages
 {
