+++ BlogTemplate/Services/SlugGenerator.cs
@@ -2,8 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using BlogTemplate.Models;
 
-namespace BlogTemplate.Models
+namespace BlogTemplate.Services
 {
     public class SlugGenerator
     {
